IOC: modified ico to create .c and .h peripheral files
Created separated peripheral files to the peripherals used for the system.

This bring better code organization and easy replacement when change from um hal to another.
